Changed listing of breweries and got rid of extra DOM Object object lists
Object object, Object object.... was rendered before each beers for
brewery table. I got rid of that by removing the ng-bind and changing
the ng-repeat to have a track by. Also, now brewery icons are displayed
when selecting a brewery instead of the name. If no logo is available,
the name is displayed.